+++ packages/angular-playground/cli/src/check-errors/reporters/xml-reporter.ts
@@ -4,13 +4,14 @@ import { ScenarioSummary } from '../verify-sandboxes';
 export class XMLReporter implements Reporter {
     constructor (
         public errors: ErrorReport[],
-        public scenarios: ScenarioSummary[]
-    ) {}
+        public scenarios: ScenarioSummary[]) {}
 
     getReport() {
         let result = `<testsuite name="Check Sandbox Errors" tests="${this.scenarios.length}" failures="${this.errors.length}" skipped="0">\n`;
         for (const scenario of this.scenarios) {
-            result += `<testcase classname="${scenario.name}" name="${scenario.description}">\n`;
+            const name = this.htmlEncode(scenario.name);
+            const description = this.htmlEncode(scenario.description);
+            result += `<testcase classname="${name}" name="${description}">\n`;
             const failure = this.errors.find(error =>
                 error.scenario === scenario.name && error.scenarioTitle === scenario.description);
             if (failure) {
@@ -24,6 +25,9 @@ export class XMLReporter implements Reporter {
     }
 
     private htmlEncode(s: string) {
+        if (!s) {
+            return '';
+        }
         return s
             .replace(/&/g, '&amp;')
             .replace(/"/g, '&quot;')