+++ packages/angular-playground/cli/src/configure.ts
@@ -1,10 +1,10 @@
-import program = require('commander');
+import * as commander from 'commander';
 import { resolve as resolvePath } from 'path';
 import { existsSync } from 'fs';
 import { REPORT_TYPE } from './error-reporter';
 
 export interface Config {
-    sourceRoot: string;
+    sourceRoots: string[];
     chunk: boolean;
     watch: boolean;
     serve: boolean;
@@ -19,16 +19,24 @@ export interface Config {
 
     angularAppName?: string;
     angularCliPath?: string;
+    angularCliHost?: string;
     angularCliPort?: number;
     angularCliAdditionalArgs?: string[];
     angularCliMaxBuffer?: number;
 }
 
+const splitCommaSeparatedList = (value) => {
+    if (!value) {
+        return ['./src/'];
+    }
+    return value.split(',');
+};
+
 export function configure(argv: any): Config {
-    program
+    commander
         .name('angular-playground')
         .option('-C, --config <path>', 'Configuration file', './angular-playground.json')
-        .option('-S, --src <path>', 'Specify component source directory', './src/')
+        .option('-S, --src <path>', 'Specify component source directories (comma separated list)', splitCommaSeparatedList)
 
         // Build options
         .option('--no-watch', 'Disable sandboxes watch', false)
@@ -46,20 +54,27 @@ export function configure(argv: any): Config {
 
         // @angular/cli options
         .option('--ng-cli-app <appName>', '@angular/cli appName')
+        .option('--ng-cli-host <ip>', '@angular/cli serve host ip', '127.0.0.1')
         .option('--ng-cli-port <n>', '@angular/cli serve port', 4201)
         .option('--ng-cli-cmd <path>', 'Path to @angular/cli executable', 'node_modules/@angular/cli/bin/ng')
         .option('--ng-cli-args <list>', 'Additional @angular/cli arguments')
         .option('--ng-cli-max-buffer <maxBuffer>', 'Specify a max buffer (for large apps)');
 
-    program.parse(argv);
-    return applyConfigurationFile(program);
+    commander.parse(argv);
+    return applyConfigurationFile(commander);
 }
 
 export function applyConfigurationFile(program: any): Config {
     const playgroundConfig = loadConfig(program.config);
+    // TODO: remove this deprecation warning at next major version
+    if (playgroundConfig.sourceRoot) {
+        console.warn('Using `sourceRoot` is deprecated. Please use `sourceRoots` instead.');
+        console.warn('See https://angularplayground.it/docs/api/configuration for more info.');
+        playgroundConfig.sourceRoots = [playgroundConfig.sourceRoot];
+    }
 
     const config: Config = {
-        sourceRoot: playgroundConfig.sourceRoot || program.src,
+        sourceRoots: playgroundConfig.sourceRoots || program.src,
         chunk: negate(playgroundConfig.noChunk) || program.chunk,
         watch: negate(playgroundConfig.noWatch) || program.watch,
         serve: negate(playgroundConfig.noServe) || program.serve,
@@ -87,6 +102,7 @@ export function applyConfigurationFile(program: any): Config {
     if (playgroundConfig.angularCli) {
         config.angularAppName = playgroundConfig.angularCli.appName || program.ngCliApp;
         config.angularCliPath = playgroundConfig.angularCli.cmdPath || program.ngCliCmd;
+        config.angularCliHost = playgroundConfig.angularCli.host || program.ngCliHost;
         config.angularCliPort = playgroundConfig.angularCli.port || program.ngCliPort;
         config.angularCliAdditionalArgs = playgroundConfig.angularCli.args || program.ngCliArgs;
         config.angularCliMaxBuffer = playgroundConfig.angularCli.maxBuffer || program.ngCliMaxBuffer;