+++ packages/angular-playground/cli/src/run.ts
@@ -1,8 +1,8 @@
+import * as getPort from 'get-port';
 import { configure, Config } from './configure';
 import { buildSandboxes } from './build-sandboxes';
 import { startWatch } from './start-watch';
 import { verifySandboxes } from './check-errors/verify-sandboxes';
-import { findFirstFreePort } from './check-errors/find-port';
 import { serveAngularCli } from './serve-angular-cli';
 import { buildAngularCli } from './build-angular-cli';
 
@@ -10,7 +10,7 @@ export async function run() {
     const config: Config = configure(process.argv);
 
     try {
-        await buildSandboxes(config.sourceRoot, config.chunk);
+        await buildSandboxes(config.sourceRoots, config.chunk);
     } catch (err) {
         throw err;
     }
@@ -20,11 +20,11 @@ export async function run() {
     }
 
     if (config.verifySandboxes) {
-        config.angularCliPort = await findFirstFreePort('127.0.0.1', 7000, 9000);
+        config.angularCliPort = await getPort({ host: config.angularCliHost });
     }
 
     if (config.watch || config.verifySandboxes) {
-        startWatch(config.sourceRoot, () => buildSandboxes(config.sourceRoot, config.chunk));
+        startWatch(config.sourceRoots, () => buildSandboxes(config.sourceRoots, config.chunk));
     }
 
     if (config.serve || config.verifySandboxes) {