+++ packages/angular-playground/cli/test/files/sandboxes-2/example6.sandbox.ts
@@ -1,0 +1,6 @@
+export default sandboxOf(ThreeDviewComponent).add(
+    'Default',
+    {
+        template: `<app-three-dview></app-three-dview>`
+    }
+);